+++ CHANGELOG.md
@@ -41,6 +41,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Links to [GENEMANIA](genemania.org)
 - Sort gene panel list on case view.
 - More automatic tests
+- Allow loading of custom annotations in VCF using the SCOUT_CUSTOM info tag.
 
 ### Fixed
 - Fix error when a gene is added to an empty dynamic gene panel